Small fixes to airbyte package
